Revert "refactor(tts): consolidate config merge"
This reverts commit b7ccba06661f52dce4ad7532f7e54fdc15fbc750.
