+++ extensions/voice-call/src/telephony-tts.ts
@@ -6,9 +6,9 @@ import {
   type SpeechProviderConfig,
   type TtsDirectiveOverrides,
 } from "openclaw/plugin-sdk/speech";
-import { mergeTtsConfigValues } from "openclaw/plugin-sdk/speech-core";
 import type { VoiceCallTtsConfig } from "./config.js";
 import type { CoreConfig } from "./core-bridge.js";
+import { deepMergeDefined } from "./deep-merge.js";
 import { convertPcmToMulaw8k } from "./telephony-audio.js";
 
 // Telephony TTS adapter that applies voice-call overrides and emits 8kHz mulaw audio.
@@ -150,7 +150,7 @@ function mergeTtsConfig(
   if (!base) {
     return override;
   }
-  return mergeTtsConfigValues(base, override) as VoiceCallTtsConfig;
+  return deepMergeDefined(base, override) as VoiceCallTtsConfig;
 }
 
 /** Resolve directive override policy for telephony synthesis. */
